clipboard patch on 0.8.3
https://st.suckless.org/patches/clipboard/st-clipboard-0.8.3.diff
---
one clipboard
The Freedesktop standard requires you to remember which clipboard you
are keeping selections in. If you switch between a terminal and browser,
you may find this UX jarring.

http://standards.freedesktop.org/clipboards-spec/clipboards-latest.txt

---
Description
This trivial patch sets CLIPBOARD on selection, the same as your
browser.

You may want to replace selpaste with clippaste in your config.h
bindings to complete the affect.